moved all diagrams to assets and fixed typos in workshop 2 cicd
