+++ src/assets/arch_diagram.png
