+++ src/assets/cloudfront-s3-diagram.png
